+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - deprecate folium module in favor of using geopandas.GeoDataFrame.explore directly
   - move plot_orientation function from bearing module to plot module
   - make matplotlib an optional dependency required only for the plot module
+  - drop pyproj package dependency
 
 ## 1.3.1.post0 (2023-05-26)